# Practica3/src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Coforge Aplication
  description: |
    Coforge Aplication
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/practices:
    post:
      tags:
        - Practices
      description: |
        Add a new practice in the system.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PracticeDTO'
      responses:
        default:
          description: Unexpected error
        '204':
          description: Practice added successfully
        '409':
          description: Error inserting practice
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Worng value for practice
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - Practices
      description: |
        Get all existing practices in the system.
      responses:
        default:
          description: Unexpected error
        '200':
          description: Practices retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PracticeDTO'
        '500':
          description: Internal server error
  /api/practices/{code}:
    get:
      tags:
        - Practices
      description: |
        Get information about a practice.
      parameters:
        - name: code
          in: path
          description: Practice code to be searched
          required: true
          schema:
            type: string
      responses:
        default:
          description: Unexpected error
        '200':
          description: Practice found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PracticeDTO'
        '404':
          description: Practice not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Practices
      description: |
        Update an existing practice in the system.
      parameters:
        - name: code
          in: path
          description: Practice name to be updated
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UPracticeDTO'
      responses:
        default:
          description: Unexpected error
        '204':
          description: Practice updated successfully
        '400':
          description: Wrong value for new practice
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Error updating the project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Practices
      description: |
        Delete an existing practice from the system.
      parameters:
        - name: code
          in: path
          description: Practice name to be deleted
          required: true
          schema:
            type: string
      responses:
        default:
          description: Unexpected error
        '204':
          description: Practice deleted successfully
        '409':
          description: Practice not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/projects:
    post:
      tags:
        - Projects
      description: |
        Add a new project in the system.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectDTO'
      responses:
        default:
          description: Unexpected error
        '204':
          description: Project added successfully
        '400':
          description: Wrong value for project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Error inserting project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - Projects
      description: |
        Get all existing projects in the system.
      responses:
        default:
          description: Unexpected error
        '200':
          description: Projects retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectDTO'
        '500':
          description: Internal server error
  /api/projects/{code}:
    get:
      tags:
        - Projects
      description: |
        Get information about a project specified by project code.
      parameters:
        - name: code
          in: path
          description: Project code to be searched
          required: true
          schema:
            type: string
      responses:
        default:
          description: Unexpected error
        '200':
          description: Project found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDTO'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Projects
      description: |
        Update an existing project in the system.
      parameters:
        - name: code
          in: path
          description: Project code to be updated
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UProjectDTO'
      responses:
        default:
          description: Unexpected error
        '204':
          description: Project updated successfully
        '400':
          description: Wrong value for new project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Error updating project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Projects
      description: |
        Delete an existing project from the system.
      parameters:
        - name: code
          in: path
          description: Project code to be deleted
          required: true
          schema:
            type: string
      responses:
        default:
          description: Unexpected error
        '204':
          description: Project deleted successfully
        '409':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    PracticeDTO:
      type: object
      properties:
        code:
          type: string
      required:
        - code
    ProjectDTO:
      type: object
      properties:
        code:
          type: string
      required:
        - code
...
    UPracticeDTO:
      type: object
      properties:
        newCode:
          type: string
    UProjectDTO:
      type: object
      properties:
        newCode:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        reason:
          type: string