# Practica3/src/main/java/com/example/practica3/DTOs/PracticeDTOs/UPracticeDTO.java
package com.example.practica3.DTOs.PracticeDTOs;


import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@AllArgsConstructor
@NoArgsConstructor
@Data
public class UPracticeDTO {
    private String newCode;
}