# Practica3/src/main/java/com/example/practica3/DTOs/ProjectDTOs/ProjectDTO.java
package com.example.practica3.DTOs.ProjectDTOs;


import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@AllArgsConstructor
@NoArgsConstructor
@Data
public class ProjectDTO {
    private String code;
}