# Practica3/src/main/java/com/example/practica3/DTOs/ProjectDTOs/UProjectDTO.java
package com.example.practica3.DTOs.ProjectDTOs;


import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@AllArgsConstructor
@NoArgsConstructor
@Data
public class UProjectDTO {
    private String newCode;
}